Remove computation for number of non-zero terms in the constraint matrix [ANT-2258] (#2496)
The goal of this PR is to avoid having to compute the number of non-zero
terms in the constraint matrix.

Some C functions require a `T*` array. Previously, we used
`std::vector<T>::data` for `CoefficientsDeLaMatriceDesContraintes` and
`IndicesColonnes`.

Instead, we use a 2-step mechanism through templated class `VectorMap`
1. Fill a `std::vector<std::pair<int, T>>` (vector of pairs)
2. Extract the vector of pairs into an `std::vector`, use it's `data`.
This `std::vector` is kept alive as long as it's `data` is used,
preventing invalid memory reads.