Remove direct use of Sirius solver (ANT-1152) (#2450)
This PR intends to be a first work in order to remove the direct call of
**Sirius** as a solver, **regarding the optimization weekly problem
only**.
We want to keep the use of **Sirius**, but only called through
**or-tools** interface.
The associated **gopro** ticket is [this
one](https://gopro-tickets.rte-france.com/browse/ANT-1152).
Here we intend to : 
- remove the use of command line option **--use-ortools**
- rename current command line option **--ortools-solver** into
**--solver**
- handle direct consequences of these changes.
- adapt legacy GUI (especially remove **use ortools** check box)

So this PR does not intend to be thorough : these first changes leave
remaining dead code to be removed
and renaming to be done. This further changes will be done in a coming
PR.

**Remaining to be done** : 
- or-tools (even if we use **Sirius** behind it) does not print
correctly MPS for named problems : optimization variables names are
composed with time steps, set once for all (0, ..., 167) before running
any year, and not updated at each week (we could expect step numbers to
be inside [0, ..., 8759] depending on the week).
So we left the code that print these correctly. We'll have to find a
solution for this.
We could leave the responsibility of printing these named MPS to Antares
code (as it did when Sirius was run alone) for a moment
- Banish all direct calls to **Sirius** from **Antares Simulator**
(calls to **Sirius** are done in **hydro heuristics**, **adequacy patch
CSR** and annual **quadratic** post-treatments of flows) does not seem
to be suitable and urgent for now.
- What else ?

---------

Co-authored-by: Florian OMNES <florian.omnes@rte-france.com>