Fix the upgrade from v8.8 to 9.2 of hydro power matrices [ANT-2742] (#2625)
Fixes problem described in issue
[ANT-2742](https://gopro-tickets.rte-france.com/browse/ANT-2742)

This fix is a quick fix : code could be cleaner. 
Some refactoring could make code more readable when reading it again in
a few months / years.
In particular, we could create an interface for reading and saving this
data (hydro power) and create 2 classes implementing this interface and
encapsulating each case (read/write daily or hourly)